chore: separate appChip and choiceChip
